renamed to useBridgeConnector
